Fixed unit test setup for aggregation controller
